+++ nav-app/src/app/app.module.ts
@@ -7,7 +7,8 @@ import 'rxjs/add/operator/map'
 import { MatInputModule } from '@angular/material/input';
 import { MatSelectModule } from '@angular/material/select';
 import { MatButtonModule } from '@angular/material/button';
-import { MatProgressSpinnerModule } from '@angular/material/progress-spinner'
+import { MatIconModule } from '@angular/material/icon';
+import { MatProgressSpinnerModule } from '@angular/material/progress-spinner';
 import { FormsModule, ReactiveFormsModule } from '@angular/forms';
 import { MatTooltipModule } from '@angular/material/tooltip';
 import { MatCheckboxModule } from '@angular/material/checkbox';
@@ -16,6 +17,7 @@ import { MatMenuModule } from '@angular/material/menu';
 import { MatExpansionModule } from '@angular/material/expansion';
 import { MatSidenavModule } from '@angular/material/sidenav';
 import { ColorPickerModule } from 'ngx-color-picker';
+import { DndModule } from 'ngx-drag-drop';
 import { PopoverModule } from 'ngx-smart-popover';
 import { HttpClientModule } from '@angular/common/http';
 
@@ -34,8 +36,13 @@ import { ContextmenuComponent } from './matrix/technique-cell/contextmenu/contex
 import { TacticCellComponent } from './matrix/tactic-cell/tactic-cell.component';
 import { VersionUpgradeComponent } from './version-upgrade/version-upgrade.component';
 import { SidebarComponent } from './sidebar/sidebar.component';
+import { LayerUpgradeComponent } from './layer-upgrade/layer-upgrade.component';
+import { ChangelogCellComponent } from './layer-upgrade/changelog-cell/changelog-cell.component';
 import { MatCardModule } from "@angular/material/card";
 import { MatDividerModule } from "@angular/material/divider";
+import { MatStepperModule } from '@angular/material/stepper';
+import { MatPaginatorModule } from '@angular/material/paginator';
+
 
 import { MarkdownModule } from "ngx-markdown";
 import { LayerInformationComponent } from './layer-information/layer-information.component';
@@ -60,6 +67,8 @@ import { SearchPopoverNotificationComponent } from './search-and-multiselect/sea
     TacticCellComponent,
     VersionUpgradeComponent,
     SidebarComponent,
+    LayerUpgradeComponent,
+    ChangelogCellComponent,
     LayerInformationComponent,
     ChangelogComponent,
     SearchPopoverNotificationComponent
@@ -73,6 +82,7 @@ import { SearchPopoverNotificationComponent } from './search-and-multiselect/sea
     ReactiveFormsModule,
     MatInputModule,
     MatButtonModule,
+    MatIconModule,
     MatProgressSpinnerModule,
     MatCheckboxModule,
     MatTooltipModule,
@@ -80,9 +90,12 @@ import { SearchPopoverNotificationComponent } from './search-and-multiselect/sea
     MatExpansionModule,
     MatDialogModule,
     ColorPickerModule,
+    DndModule,
     MatSidenavModule,
     MatCardModule,
     MatDividerModule,
+    MatStepperModule,
+    MatPaginatorModule,
     MarkdownModule.forRoot(),
     PopoverModule
   ],
@@ -90,6 +103,7 @@ import { SearchPopoverNotificationComponent } from './search-and-multiselect/sea
       MatSelectModule,
       MatInputModule,
       MatButtonModule,
+      MatIconModule,
       MatTooltipModule,
       MatMenuModule,
       MatExpansionModule,