+++ CHANGELOG.md
@@ -5,6 +5,7 @@ Version 4.4 of the Navigator restores Safari support provided you are using Safa
 
 ## New Features
 - Combined the search and multiselect interfaces into a single UI. This allows groups, software, and mitigations to be filtered alongside techniques and improves usability by moving the interface to a sidebar. See issue [#204](https://github.com/mitre-attack/attack-navigator/issues/204). 
+- Added the Layer Upgrade UI to upgrade layers created in older versions of ATT&CK. When upgrading a layer, users can view what objects have changed between the two versions and copy annotations from the previous version.  See issue [#181](https://github.com/mitre-attack/attack-navigator/issues/181).
 
 ## Improvements
 - Improved favicon for standardization with other ATT&CK tools.